Median (1-1 esque) done, Phase 2-2

+++ phase2/market.h
@@ -668,7 +668,7 @@ class market
 		dict<int, dict<order, bool>> timeExpSell;
 
 		void printname(int qty) {
-			int n = name.size() - 1;
+			int n = (int)name.size() - 1;
 			for (int i = 0; i < n; i++) {
 				cout << qty*quant[i] << " share of " << name[i] << " and ";
 			}
@@ -795,11 +795,11 @@ class market
 				}
 				else --jtr;
 			}
-			int n = bought.size();
+			int n = (int)bought.size();
 			for (int i = 0; i < n; i++) {
 				remove(bought[i], 0);
 			}
-			n = sold.size();
+			n = (int)sold.size();
 			for (int i = 0; i < n; i++) {
 				remove(sold[i], 1);
 			}
@@ -807,14 +807,14 @@ class market
 
 		bool operator == (const stock& B) const {
 			if (name.size() != B.name.size()) return false;
-			for (int i = 0; i < name.size(); i++) if (name[i] != B.name[i] || quant[i] != B.quant[i]) return false;
+			for (int i = 0; i < (int)name.size(); i++) if (name[i] != B.name[i] || quant[i] != B.quant[i]) return false;
 			return true;
 		}
 		bool operator < (const stock& B) const {
 			int i = 0;
 			while (true) {
-				if (i == name.size()) return (B.name.size() > name.size());
-				if (i == B.name.size()) return false;
+				if (i == (int)name.size()) return (B.name.size() > name.size());
+				if (i == (int)B.name.size()) return false;
 				if (name[i] < B.name[i]) return true;
 				else if (name[i] == B.name[i]) {
 					if (quant[i] == B.quant[i]) {
